requirements.txt
# ============================================================
# AI IT Helpdesk Agent - Workshop Dependencies
# ============================================================

# Core LLM & Orchestration
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LangGraph - Workflow Engine
langgraph>=0.2.0

# LangSmith - Developer Debugging & Tracing
langsmith>=0.1.0

# LangFuse - Production Observability
langfuse>=2.0.0

# Google Generative AI SDK
google-generativeai>=0.8.0

# FastAPI - REST API Layer
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0

# Rich - Beautiful terminal output for workshop demos
rich>=13.7.0

# HTTP Client
httpx>=0.25.0

# Type hints
typing-extensions>=4.8.0
